refactor: 서비스 로직 훅으로 이동
